transfering fn not related to training
