+++ src/predictions/profiles_mlcorelib/wht/pythonWHT.py
@@ -10,6 +10,7 @@ from ..utils import utils
 from ..utils.constants import TrainTablesInfo, MATERIAL_DATE_FORMAT
 from ..utils.logger import logger
 from ..connectors.Connector import Connector
+from ..trainers.MLTrainer import MLTrainer
 from .rudderPB import RudderPB
 from .mockPB import MockPB
 
@@ -589,3 +590,141 @@ class PythonWHT:
         self, model_name: str, model_hash: str, seq_no: int
     ) -> str:
         return f"{MATERIAL_PREFIX}{model_name}_{model_hash}_{seq_no:.0f}"
+
+    def check_and_generate_more_materials(
+        self,
+        get_material_func: callable,
+        materials: List[TrainTablesInfo],
+        inputs: List[dict],
+        trainer: MLTrainer,
+    ):
+        met_data_requirement = trainer.check_min_data_requirement(
+            self.connector, materials
+        )
+
+        logger.get().debug(f"Min data requirement satisfied: {met_data_requirement}")
+        if met_data_requirement or trainer.materialisation_strategy == "":
+            return materials
+
+        model_refs = [input["model_ref"] for input in inputs]
+        feature_package_path = utils.get_feature_package_path(model_refs)
+        max_materializations = (
+            trainer.materialisation_max_no_dates
+            if trainer.materialisation_strategy == "auto"
+            else len(trainer.materialisation_dates)
+        )
+
+        logger.get().info(
+            f"""Generating snapshots of past data by doing a profiles run on input models, at different points of time in the past. 
+                                Expected to run max of {2 * max_materializations} runs."""
+        )
+        for i in range(max_materializations):
+            feature_date = None
+            label_date = None
+
+            if trainer.materialisation_strategy == "auto":
+                training_dates = [
+                    utils.datetime_to_date_string(m.feature_table_date)
+                    for m in materials
+                ]
+                training_dates = [
+                    date_str for date_str in training_dates if len(date_str) != 0
+                ]
+                logger.get().info(f"training_dates : {training_dates}")
+                training_dates = sorted(
+                    training_dates,
+                    key=lambda x: datetime.strptime(x, MATERIAL_DATE_FORMAT),
+                    reverse=True,
+                )
+
+                max_feature_date = training_dates[0]
+                min_feature_date = training_dates[-1]
+
+                feature_date, label_date = utils.generate_new_training_dates(
+                    max_feature_date,
+                    min_feature_date,
+                    training_dates,
+                    trainer.prediction_horizon_days,
+                    trainer.feature_data_min_date_diff,
+                )
+                logger.get().info(
+                    f"new generated dates for feature: {feature_date}, label: {label_date}"
+                )
+            elif trainer.materialisation_strategy == "manual":
+                dates = trainer.materialisation_dates[i].split(",")
+                if len(dates) >= 2:
+                    feature_date = dates[0]
+                    label_date = dates[1]
+
+                if feature_date is None or label_date is None:
+                    continue
+
+            logger.get().info(
+                f"Looking for past data on dates {feature_date} and {label_date}. Will do a profiles run if they are not present."
+            )
+            try:
+                # Check wether the materialisation is already present
+                existing_materials = get_material_func(
+                    start_date=feature_date,
+                    end_date=feature_date,
+                    return_partial_pairs=True,
+                )
+
+                if len(existing_materials) > 0:
+                    # Check if any full sequence is exist or not
+                    complete_sequences = [
+                        sequence
+                        for sequence in existing_materials
+                        if all(element is not None for element in sequence)
+                    ]
+
+                    if len(complete_sequences) < 1:
+                        existing_material = existing_materials[0]
+                        if existing_material.feature_table_date is None:
+                            self.run(feature_package_path, feature_date)
+
+                        if existing_material.label_table_date is None:
+                            self.run(feature_package_path, label_date)
+                else:
+                    for date in [feature_date, label_date]:
+                        self.run(feature_package_path, date)
+            except Exception as e:
+                logger.get().error(str(e))
+                logger.get().error("Stopped generating new material dates.")
+                break
+
+            logger.get().info(
+                "Materialised feature and label data successfully, "
+                f"for dates {feature_date} and {label_date}"
+            )
+
+            # Get materials with new feature start date
+            # and validate min data requirement again
+
+            # For manual strategy, we need to get the materials for the selected dates separately
+            # because the selected dates may not be in the search window. so searching for materials
+            # with "feature_date" as start and end date.
+
+            # This logic will be valid for "auto" strategy as well, so we are not handling it separately.
+            materials += get_material_func(
+                start_date=feature_date, end_date=feature_date
+            )
+
+            logger.get().debug(
+                f"new feature tables: {[m.feature_table_name for m in materials]}"
+            )
+            logger.get().debug(
+                f"new label tables: {[m.label_table_name for m in materials]}"
+            )
+            if (
+                trainer.materialisation_strategy == "auto"
+                and trainer.check_min_data_requirement(self.connector, materials)
+            ):
+                logger.get().info("Minimum data requirement satisfied.")
+                break
+        if not trainer.check_min_data_requirement(self.connector, materials):
+            logger.get().error(
+                "Minimum data requirement not satisfied. Model performance may suffer. Try adding more datapoints by including more dates or increasing max_no_of_dates in the config."
+            )
+
+        return materials
