+++ tests/unit/RedshiftConnector.py
@@ -925,12 +925,11 @@ class TestCheckAndGenerateMoreMaterials(unittest.TestCase):
 
         # Call the function
         self.trainer.check_min_data_requirement = Mock(side_effect=[False, True, True])
-        result = self.trainer.check_and_generate_more_materials(
+        result = self.whtService.check_and_generate_more_materials(
             mock_get_material_func,
             materials=self.materials.copy(),
             inputs=self.inputs,
-            whtService=self.whtService,
-            connector=self.connector,
+            trainer=self.trainer,
         )
 
         # Assertions
@@ -956,12 +955,11 @@ class TestCheckAndGenerateMoreMaterials(unittest.TestCase):
         mock_rudderpb_run.side_effect = ValueError
 
         self.trainer.check_min_data_requirement = Mock(return_value=False)
-        result = self.trainer.check_and_generate_more_materials(
+        result = self.whtService.check_and_generate_more_materials(
             mock_get_material_func,
             materials=self.materials.copy(),
             inputs=self.inputs,
-            whtService=self.whtService,
-            connector=self.connector,
+            trainer=self.trainer,
         )
 
         self.assertEqual(len(result), 1)  # Only one material generated
@@ -1016,12 +1014,11 @@ class TestCheckAndGenerateMoreMaterials(unittest.TestCase):
         ]
         self.trainer.check_min_data_requirement = Mock(side_effect=[False, False, True])
         self.trainer.materialisation_strategy = "manual"
-        result = self.trainer.check_and_generate_more_materials(
+        result = self.whtService.check_and_generate_more_materials(
             mock_get_material_func,
             materials=self.materials.copy(),
             inputs=self.inputs,
-            whtService=self.whtService,
-            connector=self.connector,
+            trainer=self.trainer,
         )
 
         # Assertions
@@ -1041,12 +1038,11 @@ class TestCheckAndGenerateMoreMaterials(unittest.TestCase):
         mock_rudderpb_run.side_effect = ValueError
         self.trainer.check_min_data_requirement = Mock(return_value=False)
 
-        result = self.trainer.check_and_generate_more_materials(
+        result = self.whtService.check_and_generate_more_materials(
             mock_get_material_func,
             materials=self.materials.copy(),
             inputs=self.inputs,
-            whtService=self.whtService,
-            connector=self.connector,
+            trainer=self.trainer,
         )
 
         self.assertEqual(len(result), 1)  # Only one material generated
